Paging 3 Error

+++ app/src/main/java/com/dicoding/androidintermediate/response/Story.kt
@@ -18,7 +18,7 @@ data class Story (
     @SerializedName("createdAt")
     val createdAt: String?,
     @SerializedName("lat")
-    val lat: Float? = null,
+    val lat: Double? = null,
     @SerializedName("lon")
-    val lon: Float? = null,
+    val lon: Double? = null,
 )
